refactor(CodeGenerator) merge IProcurementPlanCodeGenerator and ProcurementPlanCodeGenerator into unified interface for code generation
