+++ Source/DakLakCoffeeSupplyChainBE/DakLakCoffeeSupplyChain.Services/Generators/ICodeGenerator.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DakLakCoffeeSupplyChain.Services.Generators
+﻿namespace DakLakCoffeeSupplyChain.Services.Generators
 {
     public interface ICodeGenerator
     {
@@ -14,6 +8,8 @@ namespace DakLakCoffeeSupplyChain.Services.Generators
 
         Task<string> GenerateBuyerCodeAsync(Guid managerId);
         Task<string> GenerateCropSeasonCodeAsync(int year);
+        Task<string> GenerateProcurementPlanCodeAsync();
+        Task<string> GenerateProcurementPlanDetailsCodeAsync();
 
     }
 }
